requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
redis
celery
kafka-python
python-multipart
